% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbox.R
\name{colorbox}
\alias{colorbox}
\title{Create a colorbox in a Markdown / Quarto Document}
\usage{
colorbox(text, color = text, maincolor = "#212529", bgcolor = "white")
}
\arguments{
\item{text}{Text to display, a character string}

\item{maincolor}{main text color in document}

\item{bgcolor}{background color in document (default: "white")}

\item{desired}{background color (e.g., \code{"red"} or \code{"#DF536B"})}
}
\value{
A character string with color-encoded text
}
\description{
Create a colorbox in a Markdown / Quarto Document
}
\examples{
# placeholder for better examples
colorbox("Gentoo", "green")
}
