% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{colorize}
\alias{colorize}
\alias{colorize_bg}
\title{Render text in color for Markdown / Quarto documents}
\usage{
colorize(text, color)

colorize_bg(text, bgcolor)
}
\arguments{
\item{text}{Text to display, a character string}

\item{color}{Color to use, a valid color designation; if missing, use \code{text} as the color}
}
\value{
A character string with color-encoded text
}
\description{
The \code{colorize()} function renders text in color for Markdown / Quarto documents. For documents to be rendered in HTML,
it uses a CSS \verb{<span>}; for documents to be converted to LaTeX and rendered as PDF, it
uses \verb{\\textcolor\{\}\{\}} from the \href{https://ctan.org/pkg/xcolor}{\code{xcolor}} package.
A companion function, \code{colorize_bg()} does the same, but uses the specified color for the background.
}
\details{
The names of colors
Note that a color name not defined in the \code{xcolor} package will trigger a LaTeX error.
e.g., \code{darkgreen} is not defined but you can use:
\verb{\\definecolor\{darkgreen\}\{RGB\}\{1,50,32\}} in a document to be rendered to PDF.

For inline text, in running text, you can use:
\preformatted{
 <span style='color: orange;'>Gentoo</span> and  <span style='color: purple;'>Adelie</span> are Penguins.
 The <span style='color: red;'>red</span> points and the <span style='color: blue;'>blue</span> points are nice
 }

In a chunk, you can also define variables with the names of colors, for ease of use:

\preformatted{
 red <- colorize('red')
 pink <- colorize("pink")
 blue <- colorize('blue')
 green <- colorize("green")
 yellow <- colorize("yellow")
 lightgreen <- colorize("lightgreen")
 darkgreen <- colorize("darkgreen")
 brown <- colorize("brown", "brown4")
 }
}
\examples{
 red <- colorize('red')
 blue <- colorize('blue')
 green <- colorize("green")
 glue::glue("There are {red} points, {blue} points and {green} points")

}
